+++ evalsense/workflow/analysers/__init__.py
@@ -1,4 +1,5 @@
 from evalsense.workflow.analysers.tabular_analyser import TabularResultAnalyser
+from evalsense.workflow.analysers.meta_result_analyser import MetaResultAnalyser
 from evalsense.workflow.analysers.metric_correlation_analyser import (
     MetricCorrelationAnalyser,
     CorrelationResults,
@@ -6,6 +7,7 @@ from evalsense.workflow.analysers.metric_correlation_analyser import (
 
 __all__ = [
     "TabularResultAnalyser",
+    "MetaResultAnalyser",
     "MetricCorrelationAnalyser",
     "CorrelationResults",
 ]